adapter: split redis components
