+++ adapter/redis_results.go
@@ -1,0 +1,21 @@
+package adapter
+
+type resultType int
+
+const (
+	resultNil resultType = iota
+	resultError
+	resultBulk
+	resultString
+	resultArray
+	resultInt
+)
+
+type redisResult struct {
+	typ     resultType
+	bulk    []byte
+	str     string
+	arr     []string
+	integer int64
+	err     error
+}
