+++ adapter/redis_transaction.go
@@ -1,0 +1,436 @@
+package adapter
+
+import (
+	"bytes"
+	"context"
+	"sort"
+	"strings"
+
+	"github.com/bootjp/elastickv/kv"
+	"github.com/bootjp/elastickv/store"
+	"github.com/cockroachdb/errors"
+	"github.com/tidwall/redcon"
+)
+
+// MULTI/EXEC/DISCARD handling
+func (r *RedisServer) multi(conn redcon.Conn, _ redcon.Command) {
+	state := getConnState(conn)
+	if state.inTxn {
+		conn.WriteError("ERR MULTI calls can not be nested")
+		return
+	}
+	state.inTxn = true
+	state.queue = nil
+	conn.WriteString("OK")
+}
+
+func (r *RedisServer) discard(conn redcon.Conn, _ redcon.Command) {
+	state := getConnState(conn)
+	if !state.inTxn {
+		conn.WriteError("ERR DISCARD without MULTI")
+		return
+	}
+	state.inTxn = false
+	state.queue = nil
+	conn.WriteString("OK")
+}
+
+func (r *RedisServer) exec(conn redcon.Conn, _ redcon.Command) {
+	state := getConnState(conn)
+	if !state.inTxn {
+		conn.WriteError("ERR EXEC without MULTI")
+		return
+	}
+
+	if !r.coordinator.IsLeader() {
+		if err := r.proxyExec(conn, state.queue); err != nil {
+			conn.WriteError(err.Error())
+		}
+		state.inTxn = false
+		state.queue = nil
+		return
+	}
+
+	results, err := r.runTransaction(state.queue)
+	state.inTxn = false
+	state.queue = nil
+	if err != nil {
+		conn.WriteError(err.Error())
+		return
+	}
+
+	r.writeResults(conn, results)
+}
+
+type txnValue struct {
+	raw     []byte
+	deleted bool
+	dirty   bool
+	loaded  bool
+}
+
+type txnContext struct {
+	server     *RedisServer
+	working    map[string]*txnValue
+	listStates map[string]*listTxnState
+	startTS    uint64
+}
+
+type listTxnState struct {
+	meta       store.ListMeta
+	metaExists bool
+	appends    [][]byte
+	deleted    bool
+}
+
+func (t *txnContext) load(key []byte) (*txnValue, error) {
+	k := string(key)
+	if tv, ok := t.working[k]; ok {
+		return tv, nil
+	}
+	tv := &txnValue{}
+	val, err := t.server.readValueAt(key, t.startTS)
+	if err != nil && !errors.Is(err, store.ErrKeyNotFound) {
+		return nil, errors.WithStack(err)
+	}
+	tv.raw = val
+	tv.loaded = true
+	t.working[k] = tv
+	return tv, nil
+}
+
+func (t *txnContext) loadListState(key []byte) (*listTxnState, error) {
+	k := string(key)
+	if st, ok := t.listStates[k]; ok {
+		return st, nil
+	}
+
+	meta, exists, err := t.server.loadListMetaAt(context.Background(), key, t.startTS)
+	if err != nil {
+		return nil, err
+	}
+
+	st := &listTxnState{
+		meta:       meta,
+		metaExists: exists,
+		appends:    [][]byte{},
+	}
+	t.listStates[k] = st
+	return st, nil
+}
+
+func (t *txnContext) listLength(st *listTxnState) int64 {
+	return st.meta.Len + int64(len(st.appends))
+}
+
+func (t *txnContext) apply(cmd redcon.Command) (redisResult, error) {
+	switch strings.ToUpper(string(cmd.Args[0])) {
+	case "SET":
+		return t.applySet(cmd)
+	case "DEL":
+		return t.applyDel(cmd)
+	case "GET":
+		return t.applyGet(cmd)
+	case "EXISTS":
+		return t.applyExists(cmd)
+	case "RPUSH":
+		return t.applyRPush(cmd)
+	case "LRANGE":
+		return t.applyLRange(cmd)
+	default:
+		return redisResult{}, errors.WithStack(errors.Newf("ERR unsupported command '%s'", cmd.Args[0]))
+	}
+}
+
+func (t *txnContext) applySet(cmd redcon.Command) (redisResult, error) {
+	if isList, err := t.server.isListKeyAt(context.Background(), cmd.Args[1], t.startTS); err != nil {
+		return redisResult{}, err
+	} else if isList {
+		return redisResult{typ: resultError, err: errors.New("WRONGTYPE Operation against a key holding the wrong kind of value")}, nil
+	}
+
+	tv, err := t.load(cmd.Args[1])
+	if err != nil {
+		return redisResult{}, err
+	}
+	tv.raw = cmd.Args[2]
+	tv.deleted = false
+	tv.dirty = true
+	return redisResult{typ: resultString, str: "OK"}, nil
+}
+
+func (t *txnContext) applyDel(cmd redcon.Command) (redisResult, error) {
+	// handle list delete separately
+	if isList, err := t.server.isListKeyAt(context.Background(), cmd.Args[1], t.startTS); err != nil {
+		return redisResult{}, err
+	} else if isList {
+		st, err := t.loadListState(cmd.Args[1])
+		if err != nil {
+			return redisResult{}, err
+		}
+		st.deleted = true
+		st.appends = nil
+		return redisResult{typ: resultInt, integer: 1}, nil
+	}
+
+	tv, err := t.load(cmd.Args[1])
+	if err != nil {
+		return redisResult{}, err
+	}
+	tv.deleted = true
+	tv.dirty = true
+	return redisResult{typ: resultInt, integer: 1}, nil
+}
+
+func (t *txnContext) applyGet(cmd redcon.Command) (redisResult, error) {
+	if isList, err := t.server.isListKeyAt(context.Background(), cmd.Args[1], t.startTS); err != nil {
+		return redisResult{}, err
+	} else if isList {
+		return redisResult{typ: resultError, err: errors.New("WRONGTYPE Operation against a key holding the wrong kind of value")}, nil
+	}
+
+	tv, err := t.load(cmd.Args[1])
+	if err != nil {
+		return redisResult{}, err
+	}
+	if tv.deleted || tv.raw == nil {
+		return redisResult{typ: resultNil}, nil
+	}
+	return redisResult{typ: resultBulk, bulk: tv.raw}, nil
+}
+
+func (t *txnContext) applyExists(cmd redcon.Command) (redisResult, error) {
+	if isList, err := t.server.isListKeyAt(context.Background(), cmd.Args[1], t.startTS); err != nil {
+		return redisResult{}, err
+	} else if isList {
+		return redisResult{typ: resultInt, integer: 1}, nil
+	}
+
+	tv, err := t.load(cmd.Args[1])
+	if err != nil {
+		return redisResult{}, err
+	}
+	if tv.deleted || tv.raw == nil {
+		return redisResult{typ: resultInt, integer: 0}, nil
+	}
+	return redisResult{typ: resultInt, integer: 1}, nil
+}
+
+func (t *txnContext) applyRPush(cmd redcon.Command) (redisResult, error) {
+	st, err := t.loadListState(cmd.Args[1])
+	if err != nil {
+		return redisResult{}, err
+	}
+
+	for _, v := range cmd.Args[2:] {
+		st.appends = append(st.appends, bytes.Clone(v))
+	}
+
+	return redisResult{typ: resultInt, integer: t.listLength(st)}, nil
+}
+
+func (t *txnContext) applyLRange(cmd redcon.Command) (redisResult, error) {
+	st, err := t.loadListState(cmd.Args[1])
+	if err != nil {
+		return redisResult{}, err
+	}
+
+	s, e, err := parseRangeBounds(cmd.Args[2], cmd.Args[3], int(t.listLength(st)))
+	if err != nil {
+		return redisResult{}, err
+	}
+	if e < s {
+		return redisResult{typ: resultArray, arr: []string{}}, nil
+	}
+
+	out, err := t.listRangeValues(cmd.Args[1], st, s, e)
+	if err != nil {
+		return redisResult{}, err
+	}
+
+	return redisResult{typ: resultArray, arr: out}, nil
+}
+
+func parseRangeBounds(startRaw, endRaw []byte, total int) (int, int, error) {
+	start, err := parseInt(startRaw)
+	if err != nil {
+		return 0, 0, err
+	}
+	end, err := parseInt(endRaw)
+	if err != nil {
+		return 0, 0, err
+	}
+	s, e := clampRange(start, end, total)
+	return s, e, nil
+}
+
+func (t *txnContext) listRangeValues(key []byte, st *listTxnState, s, e int) ([]string, error) {
+	persistedLen := int(st.meta.Len)
+
+	switch {
+	case e < persistedLen:
+		return t.server.fetchListRange(context.Background(), key, st.meta, int64(s), int64(e), t.startTS)
+	case s >= persistedLen:
+		return appendValues(st.appends, s-persistedLen, e-persistedLen), nil
+	default:
+		head, err := t.server.fetchListRange(context.Background(), key, st.meta, int64(s), int64(persistedLen-1), t.startTS)
+		if err != nil {
+			return nil, err
+		}
+		tail := appendValues(st.appends, 0, e-persistedLen)
+		return append(head, tail...), nil
+	}
+}
+
+func appendValues(buf [][]byte, start, end int) []string {
+	out := make([]string, 0, end-start+1)
+	for i := start; i <= end; i++ {
+		out = append(out, string(buf[i]))
+	}
+	return out
+}
+
+func (t *txnContext) commit() error {
+	elems := t.buildKeyElems()
+
+	listElems, err := t.buildListElems()
+	if err != nil {
+		return err
+	}
+
+	elems = append(elems, listElems...)
+	if len(elems) == 0 {
+		return nil
+	}
+
+	group := &kv.OperationGroup[kv.OP]{IsTxn: true, Elems: elems, StartTS: t.startTS}
+	if _, err := t.server.coordinator.Dispatch(group); err != nil {
+		return errors.WithStack(err)
+	}
+	return nil
+}
+
+func (t *txnContext) buildKeyElems() []*kv.Elem[kv.OP] {
+	keys := make([]string, 0, len(t.working))
+	for k := range t.working {
+		keys = append(keys, k)
+	}
+	sort.Strings(keys)
+
+	var elems []*kv.Elem[kv.OP]
+	for _, k := range keys {
+		tv := t.working[k]
+		if !tv.dirty {
+			continue
+		}
+		key := []byte(k)
+		if tv.deleted {
+			elems = append(elems, &kv.Elem[kv.OP]{Op: kv.Del, Key: key})
+			continue
+		}
+		elems = append(elems, &kv.Elem[kv.OP]{Op: kv.Put, Key: key, Value: tv.raw})
+	}
+	return elems
+}
+
+func (t *txnContext) buildListElems() ([]*kv.Elem[kv.OP], error) {
+	listKeys := make([]string, 0, len(t.listStates))
+	for k := range t.listStates {
+		listKeys = append(listKeys, k)
+	}
+	sort.Strings(listKeys)
+
+	var elems []*kv.Elem[kv.OP]
+	for _, k := range listKeys {
+		st := t.listStates[k]
+		userKey := []byte(k)
+
+		if st.deleted {
+			// delete all persisted list items
+			for seq := st.meta.Head; seq < st.meta.Tail; seq++ {
+				elems = append(elems, &kv.Elem[kv.OP]{Op: kv.Del, Key: listItemKey(userKey, seq)})
+			}
+			elems = append(elems, &kv.Elem[kv.OP]{Op: kv.Del, Key: listMetaKey(userKey)})
+			continue
+		}
+		if len(st.appends) == 0 {
+			continue
+		}
+
+		startSeq := st.meta.Head + st.meta.Len
+		for i, v := range st.appends {
+			elems = append(elems, &kv.Elem[kv.OP]{
+				Op:    kv.Put,
+				Key:   listItemKey(userKey, startSeq+int64(i)),
+				Value: v,
+			})
+		}
+
+		st.meta.Len += int64(len(st.appends))
+		st.meta.Tail = st.meta.Head + st.meta.Len
+		metaBytes, err := store.MarshalListMeta(st.meta)
+		if err != nil {
+			return nil, errors.WithStack(err)
+		}
+		elems = append(elems, &kv.Elem[kv.OP]{Op: kv.Put, Key: listMetaKey(userKey), Value: metaBytes})
+	}
+	return elems, nil
+}
+
+func (r *RedisServer) runTransaction(queue []redcon.Command) ([]redisResult, error) {
+	if err := r.coordinator.VerifyLeader(); err != nil {
+		return nil, errors.WithStack(err)
+	}
+
+	startTS := r.coordinator.Clock().Next()
+	if last := r.store.LastCommitTS(); last > startTS {
+		startTS = last
+	}
+
+	ctx := &txnContext{
+		server:     r,
+		working:    map[string]*txnValue{},
+		listStates: map[string]*listTxnState{},
+		startTS:    startTS,
+	}
+
+	results := make([]redisResult, 0, len(queue))
+	for _, cmd := range queue {
+		res, err := ctx.apply(cmd)
+		if err != nil {
+			return nil, err
+		}
+		results = append(results, res)
+	}
+
+	if err := ctx.commit(); err != nil {
+		return nil, err
+	}
+
+	return results, nil
+}
+
+func (r *RedisServer) writeResults(conn redcon.Conn, results []redisResult) {
+	conn.WriteArray(len(results))
+	for _, res := range results {
+		switch res.typ {
+		case resultNil:
+			conn.WriteNull()
+		case resultError:
+			conn.WriteError(res.err.Error())
+		case resultBulk:
+			conn.WriteBulk(res.bulk)
+		case resultString:
+			conn.WriteString(res.str)
+		case resultArray:
+			conn.WriteArray(len(res.arr))
+			for _, s := range res.arr {
+				conn.WriteBulkString(s)
+			}
+		case resultInt:
+			conn.WriteInt64(res.integer)
+		default:
+			conn.WriteNull()
+		}
+	}
+}
