+++ adapter/redis_utils.go
@@ -1,0 +1,31 @@
+package adapter
+
+import (
+	"strconv"
+
+	"github.com/cockroachdb/errors"
+)
+
+func parseInt(b []byte) (int, error) {
+	i, err := strconv.Atoi(string(b))
+	return i, errors.WithStack(err)
+}
+
+func clampRange(start, end, length int) (int, int) {
+	if start < 0 {
+		start = length + start
+	}
+	if end < 0 {
+		end = length + end
+	}
+	if start < 0 {
+		start = 0
+	}
+	if end >= length {
+		end = length - 1
+	}
+	if end < start {
+		return 0, -1
+	}
+	return start, end
+}
